Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int radius;
+    float radius;
     float Area;
     Area = (radius * radius) * 3.14;
-    scanf("%d", &radius);
+    scanf("%f", &radius);
     printf("Area: %.2f",Area);
     return 0;
 }
